refactoring alarm manager
