+++ app/src/main/java/com/emikhalets/mydates/foreground/RebootReceiver.kt
@@ -3,16 +3,13 @@ package com.emikhalets.mydates.foreground
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.emikhalets.mydates.utils.setEventAlarm
-import com.emikhalets.mydates.utils.setUpdatingAlarm
-
+import com.emikhalets.mydates.utils.AppAlarmManager
 
 class RebootReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action == Intent.ACTION_BOOT_COMPLETED) {
-            context.setUpdatingAlarm()
-            context.setEventAlarm()
+            AppAlarmManager.scheduleAllAlarms(context)
         }
     }
 }
