+++ app/src/main/java/com/emikhalets/mydates/foreground/UpdateEventsReceiver.kt
@@ -5,12 +5,12 @@ import android.content.Context
 import android.content.Intent
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
-import com.emikhalets.mydates.utils.setUpdatingAlarm
+import com.emikhalets.mydates.utils.AppAlarmManager
 
 class UpdateEventsReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context, intent: Intent) {
-        context.setUpdatingAlarm()
+        AppAlarmManager.scheduleUpdatingAlarm(context)
         val worker = OneTimeWorkRequestBuilder<UpdateEventsWorker>().build()
         WorkManager.getInstance(context).enqueue(worker)
     }
